+++ androidx-kt/src/main/java/me/panpf/androidxkt/content/Intent.kt
@@ -26,6 +26,7 @@ import android.net.Uri
 import android.provider.MediaStore
 import android.provider.Settings
 import android.support.annotation.RequiresPermission
+import android.support.v4.app.Fragment
 
 
 /**
@@ -184,6 +185,16 @@ fun Context.canStartActivity(intent: Intent): Boolean {
 }
 
 /**
+ * Test if you can start Activity
+ */
+fun Fragment.canStartActivity(intent: Intent): Boolean {
+    val activity = this.activity
+            ?: throw IllegalStateException("Fragment $this not attached to Activity")
+
+    return activity.canStartActivity(intent)
+}
+
+/**
  * Safely launch an Activity, catch ActivityNotFoundException and return false
  */
 fun Context.safeStartActivity(intent: Intent): Boolean {
@@ -198,6 +209,16 @@ fun Context.safeStartActivity(intent: Intent): Boolean {
         e.printStackTrace()
         false
     }
+}
+
+/**
+ * Safely launch an Activity, catch ActivityNotFoundException and return false
+ */
+fun Fragment.safeStartActivity(intent: Intent): Boolean {
+    val activity = this.activity
+            ?: throw IllegalStateException("Fragment $this not attached to Activity")
+
+    return activity.safeStartActivity(intent)
 }
 
 /**